fix type error

+++ project_opt1_quant/SW_Quantization/src/fpga_api_on_cpu.cpp
@@ -26,8 +26,8 @@ FPGA::FPGA(off_t data_addr, off_t output_addr, int m_size, int v_size)
   qout_ = new int[m_size_];
   qout_M = new int[v_size_*v_size_];
 
-  output_ = new unsigned int[m_size_]; // use output_ as tempolar output
-  output_M = new unsigned int[v_size_*v_size_]; // use output_M as tempolar output
+  output_ = new float[m_size_]; // use output_ as tempolar output
+  output_M = new float[v_size_*v_size_]; // use output_M as tempolar output
 
   data_ = new float[data_size_];
   data_M = new float[data_size_M];
